Move donation pop up to root
